pyproject.toml
[project]
name = "2025-4chan-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "pyfaktory>=0.2.10",
    "ruff>=0.13.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
